added News componenet and relative scss
