--- src/scss/News-Spotlights.scss
@@ -1,88 +1,0 @@
-.news-spotlights-container {
-    display: flex;
-    flex-direction: row;
-    background: #E5E5E5;
-    align-items: center;
-    justify-content: center;
-
-    .cards-wrapper {
-        display: flex;
-        flex-direction: row;
-        margin: 87px auto 108px auto;
-        flex-wrap: wrap;
-        justify-content: center;
-    }
-
-    .card__news,
-    .card__spotlight {
-        width: 480px;
-        height: 578px;
-        margin-right: 20px;
-        padding-left: 19px;
-        @include card-style;
-
-        h2 {
-            margin-top: 35px;
-            @include header-two;
-            line-height: 36px;
-            letter-spacing: -0.03em;
-            color: $openLawArapawa;
-            width: 205px;
-            height: 35px;
-        }
-
-        .picture {
-            margin-left: 1px;
-            width: 438px;
-            height: 216px;
-            background: #C4C4C4;
-        }
-
-        h3 {
-            @include header-four;
-            line-height: 24px;
-            width: 205px;
-            height: 28px;
-            margin-top: 30px;
-            margin-bottom: 5px;
-            color: $openLawArapawa;
-        }
-
-        p {
-            @include body;
-            width: 439px;
-            height: 108px;
-            line-height: 136.7%;
-            /* or 21px */
-            color: $openLawArapawa;
-            margin-bottom: 0;
-        }
-
-        .link {
-            @include body;
-            width: 102px;
-            line-height: 18px;
-            color: $openLawTeal;
-            font-weight: 900;
-            text-decoration-color: $openLawTeal;
-        }
-
-    }
-
-    .card__spotlight {
-        margin-right: 0px;
-    }
-
-
-}
-
-@media (max-width: 979px) {
-    .card__news {
-        margin: 0 0 20px 0px !important;
-    }
-
-    .cards-wrapper {
-        margin: 35px auto 35px auto !important;
-        flex-direction: column !important;
-    }
-}
