+++ src/scss/News.scss
@@ -1,0 +1,74 @@
+.news-container {
+    background: $newsBackgroundColor;
+    height: 562px;
+
+    .news-items {
+        padding: 46px 73px 71px 71px;
+        @include card-style;
+        position: relative;
+        top: -80px;
+        display: flex;
+        margin: 0 auto;
+        flex-direction: column;
+        width: 980px;
+        height: 538px;
+
+        h1 {
+            @include header-one;
+            line-height: 48px;
+            color: $openLawArapawa;
+            width: 205px;
+            margin-bottom: 43px;
+        }
+
+        .news-cards-wrapper {
+            display: flex;
+            flex-direction: row;
+
+            justify-content: space-between;
+
+            .news-card {
+                width: 266px;
+
+                .picture {
+                    margin-left: 1px;
+                    height: 181px;
+                    background: #C4C4C4;
+                }
+
+                h2 {
+                    @include header-four;
+                    line-height: 24px;
+                    width: 205px;
+                    height: 28px;
+                    margin-top: 22px;
+                    color: $openLawArapawa;
+                }
+
+                p {
+                    @include body;
+                    height: 84px;
+                    line-height: 136.7%;
+                    /* or 21px */
+                    color: $openLawArapawa;
+                    margin-bottom: 0;
+                }
+
+                .link {
+                    @include body;
+                    width: 102px;
+                    height: 13px;
+                    line-height: 18px;
+                    color: $openLawTeal;
+                    margin-top: 15px;
+                    text-decoration-color: $openLawTeal;
+                }
+            }
+
+        }
+
+    }
+
+
+
+}
